create move cemetery to deck
