+++ client/src/battle/deck.ts
@@ -1,4 +1,5 @@
 import { CardBaseType, CardType } from '../types/model/index'
+import { deckShuffle } from '../common/battle'
 
 export const initializeDeck = (cardList: CardBaseType[]): CardType[] => {
   let defaultDeck: CardType[] = []
@@ -39,14 +40,4 @@ export const initializeDeck = (cardList: CardBaseType[]): CardType[] => {
   }
   defaultDeck = deckShuffle(defaultDeck)
   return defaultDeck
-}
-
-const deckShuffle = (deck: CardType[]): CardType[] => {
-  for (let i = deck.length - 1; i > 0; i--) {
-    const r = Math.floor(Math.random() * (i + 1))
-    const tmp = deck[i]
-    deck[i] = deck[r]
-    deck[r] = tmp
-  }
-  return deck
 }
